backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sentence-transformers==3.4.1
Pillow==11.2.1
